requirements.txt
discord.py >1.3, <1.4
aiohttp >3.6, <4
aiofiles >=0.5, <0.6

